IS-1187 added campaigns for nimble characters
