+++ app/javascript/applications/CharKeeperApp/pages/Content/Character/Daggerheart.jsx
@@ -1,31 +1,28 @@
 import { createSignal, createMemo, Show, Switch, Match } from 'solid-js';
-import * as i18n from '@solid-primitives/i18n';
 import { createWindowSize } from '@solid-primitives/resize-observer';
 
 import {
   DaggerheartTraits, DaggerheartStatic, DaggerheartHealth, DaggerheartBeastform, DaggerheartCompanion,
   DaggerheartDomainCards, DaggerheartRest, DaggerheartLeveling, DaggerheartExperience, DaggerheartTransform,
-  DaggerheartStances, DaggerheartBonuses, DaggerheartCraft, DaggerheartScars, DaggerheartInfo, DaggerheartLootTable
+  DaggerheartStances, DaggerheartBonuses, DaggerheartCraft, DaggerheartScars, DaggerheartInfo, DaggerheartLootTable,
+  DaggerheartEquipment
 } from '../../../pages';
 import {
-  CharacterNavigation, Equipment, Notes, Avatar, ContentWrapper, Feats, createRoll, Conditions, Combat, Gold
+  CharacterNavigation, Notes, Avatar, ContentWrapper, Feats, createRoll, Conditions, Combat, Gold
 } from '../../../components';
 import { useAppLocale } from '../../../context';
 import { localize } from '../../../helpers';
 
 const TRANSLATION = {
   en: {
-    helpMessage: "You can choose either a two-handed primary weapon, or a one-handed primary weapon and a one-handed secondary weapon, then equip them. You can choose one set of armor and equip it. You can choose any other items.",
     domainHelpMessage: "To start, look at all the level 1 cards from your class's two domains and choose two cards. You can take one from each domain or two from a single domain.",
     levelingHelpMessage: "In the future on this tab you can level up your character."
   },
   ru: {
-    helpMessage: "Вы можете выбрать двуручное оружие или одноручное основное и одноручное дополнительное оружие, а затем экипировать его. Вы можете выбрать набор брони и экипировать его. Вы также можете выбрать другие вещи.",
     domainHelpMessage: "В начале изучите все карты 1 уровня из двух доменов вашего класса и выберите 2 карты. Вы можете выбрать по карте из каждого домена или обе карты из одного домена.",
     levelingHelpMessage: "В будущем на этой вкладке вы сможете указывать уровень вашего персонажа."
   },
   es: {
-    helpMessage: "Puedes elegir un arma principal de dos manos, o un arma principal de una mano y un arma secundaria de una mano, luego equiparlos. Puedes elegir un conjunto de armadura y equiparlo. También puedes elegir otros objetos.",
     domainHelpMessage: "Para empezar, mira todas las cartas de nivel 1 de los dos dominios de tu clase y elige dos cartas. Puedes elegir una de cada dominio o dos del mismo dominio.",
     levelingHelpMessage: "En el futuro en esta pestaña podrás subir de nivel a tu personaje."
   }
@@ -39,30 +36,7 @@ export const Daggerheart = (props) => {
   const [activeTab, setActiveTab] = createSignal('combat');
 
   const { Roll, openDualityTest, openDualityAttack, openDices } = createRoll();
-  const [locale, dict] = useAppLocale();
-
-  const t = i18n.translator(dict);
-
-  const primaryWeaponFilterT1 = (item) => item.kind === 'primary weapon' && item.info.tier === 1 && item.info.damage_type === 'physical';
-  const primaryWeaponFilterT1Magic = (item) => item.kind === 'primary weapon' && item.info.tier === 1 && item.info.damage_type === 'magic';
-  const primaryWeaponFilterT2 = (item) => item.kind === 'primary weapon' && item.info.tier === 2 && item.info.damage_type === 'physical';
-  const primaryWeaponFilterT2Magic = (item) => item.kind === 'primary weapon' && item.info.tier === 2 && item.info.damage_type === 'magic';
-  const primaryWeaponFilterT3 = (item) => item.kind === 'primary weapon' && item.info.tier === 3 && item.info.damage_type === 'physical';
-  const primaryWeaponFilterT3Magic = (item) => item.kind === 'primary weapon' && item.info.tier === 3 && item.info.damage_type === 'magic';
-  const primaryWeaponFilterT4 = (item) => item.kind === 'primary weapon' && item.info.tier === 4 && item.info.damage_type === 'physical';
-  const primaryWeaponFilterT4Magic = (item) => item.kind === 'primary weapon' && item.info.tier === 4 && item.info.damage_type === 'magic';
-  const secondaryWeaponFilterT1 = (item) => item.kind === 'secondary weapon' && item.info.tier === 1;
-  const secondaryWeaponFilterT2 = (item) => item.kind === 'secondary weapon' && item.info.tier === 2;
-  const secondaryWeaponFilterT3 = (item) => item.kind === 'secondary weapon' && item.info.tier === 3;
-  const secondaryWeaponFilterT4 = (item) => item.kind === 'secondary weapon' && item.info.tier === 4;
-  const armorFilterT1 = (item) => item.kind === 'armor' && item.info.tier === 1;
-  const armorFilterT2 = (item) => item.kind === 'armor' && item.info.tier === 2;
-  const armorFilterT3 = (item) => item.kind === 'armor' && item.info.tier === 3;
-  const armorFilterT4 = (item) => item.kind === 'armor' && item.info.tier === 4;
-  const itemsFilter = (item) => item.kind === 'item';
-  const consumablesFilter = (item) => item.kind === 'consumables';
-  const recipesFilter = (item) => item.kind === 'recipe';
-  const upgradesFilter = (item) => item.kind === 'upgrade';
+  const [locale] = useAppLocale();
 
   const ancestryFilter = (item) => item.origin === 'ancestry';
   const communityFilter = (item) => item.origin === 'community';
@@ -169,41 +143,18 @@ export const Daggerheart = (props) => {
               </div>
             </Match>
             <Match when={activeMobileTab() === 'equipment'}>
-              <Equipment
+              <DaggerheartEquipment
                 upgrades={['primary weapon', 'secondary weapon', 'armor']}
                 character={character()}
-                itemFilters={[
-                  { title: `${t('equipment.primaryWeapon')} T1`, callback: primaryWeaponFilterT1 },
-                  { title: `${t('equipment.primaryWeaponMagic')} T1`, callback: primaryWeaponFilterT1Magic },
-                  { title: `${t('equipment.primaryWeapon')} T2`, callback: primaryWeaponFilterT2 },
-                  { title: `${t('equipment.primaryWeaponMagic')} T2`, callback: primaryWeaponFilterT2Magic },
-                  { title: `${t('equipment.primaryWeapon')} T3`, callback: primaryWeaponFilterT3 },
-                  { title: `${t('equipment.primaryWeaponMagic')} T3`, callback: primaryWeaponFilterT3Magic },
-                  { title: `${t('equipment.primaryWeapon')} T4`, callback: primaryWeaponFilterT4 },
-                  { title: `${t('equipment.primaryWeaponMagic')} T4`, callback: primaryWeaponFilterT4Magic },
-                  { title: `${t('equipment.secondaryWeapon')} T1`, callback: secondaryWeaponFilterT1 },
-                  { title: `${t('equipment.secondaryWeapon')} T2`, callback: secondaryWeaponFilterT2 },
-                  { title: `${t('equipment.secondaryWeapon')} T3`, callback: secondaryWeaponFilterT3 },
-                  { title: `${t('equipment.secondaryWeapon')} T4`, callback: secondaryWeaponFilterT4 },
-                  { title: `${t('equipment.armorList')} T1`, callback: armorFilterT1 },
-                  { title: `${t('equipment.armorList')} T2`, callback: armorFilterT2 },
-                  { title: `${t('equipment.armorList')} T3`, callback: armorFilterT3 },
-                  { title: `${t('equipment.armorList')} T4`, callback: armorFilterT4 },
-                  { title: t('equipment.itemsList'), callback: itemsFilter },
-                  { title: t('equipment.consumables'), callback: consumablesFilter },
-                  { title: t('equipment.recipes'), callback: recipesFilter },
-                  { title: t('equipment.upgrades'), callback: upgradesFilter }
-                ]}
                 onReplaceCharacter={props.onReplaceCharacter}
                 onReloadCharacter={props.onReloadCharacter}
                 currentGuideStep={character().guide_step}
                 guideStep={3}
-                helpMessage={localize(TRANSLATION, locale())['helpMessage']}
                 onNextGuideStepClick={() => setActiveMobileTab('domainCards')}
                 lootTableComponent={DaggerheartLootTable}
               >
                 <Gold character={character()} onReplaceCharacter={props.onReplaceCharacter} />
-              </Equipment>
+              </DaggerheartEquipment>
             </Match>
             <Match when={activeMobileTab() === 'domainCards'}>
               <DaggerheartDomainCards
@@ -336,41 +287,18 @@ export const Daggerheart = (props) => {
               </div>
             </Match>
             <Match when={activeTab() === 'equipment'}>
-              <Equipment
+              <DaggerheartEquipment
                 upgrades={['primary weapon', 'secondary weapon', 'armor']}
                 character={character()}
-                itemFilters={[
-                  { title: `${t('equipment.primaryWeapon')} T1`, callback: primaryWeaponFilterT1 },
-                  { title: `${t('equipment.primaryWeaponMagic')} T1`, callback: primaryWeaponFilterT1Magic },
-                  { title: `${t('equipment.primaryWeapon')} T2`, callback: primaryWeaponFilterT2 },
-                  { title: `${t('equipment.primaryWeaponMagic')} T2`, callback: primaryWeaponFilterT2Magic },
-                  { title: `${t('equipment.primaryWeapon')} T3`, callback: primaryWeaponFilterT3 },
-                  { title: `${t('equipment.primaryWeaponMagic')} T3`, callback: primaryWeaponFilterT3Magic },
-                  { title: `${t('equipment.primaryWeapon')} T4`, callback: primaryWeaponFilterT4 },
-                  { title: `${t('equipment.primaryWeaponMagic')} T4`, callback: primaryWeaponFilterT4Magic },
-                  { title: `${t('equipment.secondaryWeapon')} T1`, callback: secondaryWeaponFilterT1 },
-                  { title: `${t('equipment.secondaryWeapon')} T2`, callback: secondaryWeaponFilterT2 },
-                  { title: `${t('equipment.secondaryWeapon')} T3`, callback: secondaryWeaponFilterT3 },
-                  { title: `${t('equipment.secondaryWeapon')} T4`, callback: secondaryWeaponFilterT4 },
-                  { title: `${t('equipment.armorList')} T1`, callback: armorFilterT1 },
-                  { title: `${t('equipment.armorList')} T2`, callback: armorFilterT2 },
-                  { title: `${t('equipment.armorList')} T3`, callback: armorFilterT3 },
-                  { title: `${t('equipment.armorList')} T4`, callback: armorFilterT4 },
-                  { title: t('equipment.itemsList'), callback: itemsFilter },
-                  { title: t('equipment.consumables'), callback: consumablesFilter },
-                  { title: t('equipment.recipes'), callback: recipesFilter },
-                  { title: t('equipment.upgrades'), callback: upgradesFilter }
-                ]}
                 onReplaceCharacter={props.onReplaceCharacter}
                 onReloadCharacter={props.onReloadCharacter}
                 currentGuideStep={character().guide_step}
                 guideStep={3}
-                helpMessage={localize(TRANSLATION, locale())['helpMessage']}
-                onNextGuideStepClick={() => setActiveTab('domainCards')}
+                onNextGuideStepClick={() => setActiveMobileTab('domainCards')}
                 lootTableComponent={DaggerheartLootTable}
               >
                 <Gold character={character()} onReplaceCharacter={props.onReplaceCharacter} />
-              </Equipment>
+              </DaggerheartEquipment>
             </Match>
             <Match when={activeTab() === 'domainCards'}>
               <DaggerheartDomainCards
