+++ app/javascript/applications/CharKeeperApp/pages/Content/Character/Nimble/Skills.jsx
@@ -134,7 +134,7 @@ export const NimbleSkills = (props) => {
                               width="28"
                               height="28"
                               text={modifier(skill.modifier)}
-                              onClick={() => props.openD20Test(`/check skill ${slug}`, skill.name, skill.modifier)}
+                              onClick={() => props.openD20Test(`/check skill ${skill.slug}`, skill.name, skill.modifier)}
                             />
                           </div>
                         }
