+++ app/services/bot_context_v2/commands/checks/nimble_attack.rb
@@ -32,7 +32,8 @@ module BotContextV2
           {
             status: 'success',
             total: primary_roll + secondary_rolls.sum + crit_rolls.sum + values[:bonus],
-            rolls: [primary_roll] + secondary_rolls + crit_rolls
+            rolls: [primary_roll] + secondary_rolls + crit_rolls,
+            bonus: values[:bonus]
           }
         end
