+++ app/views/bots_v2/character_bot/check/dnd/attack.text.erb
@@ -1,2 +1,2 @@
 <b><%= character.name %></b> <%= I18n.t('services.bot_context.representers.check.dnd.attack', target: target) %>
-<%= I18n.t('services.bot_context.representers.check.result', result: dnd_roll_result_message(result)) %>
+<%= I18n.t('services.bot_context.representers.check.result') %><%= dnd_roll_result_message(result) %>
