+++ app/views/bots_v2/character_bot/check/fate/skill.text.erb
@@ -1,2 +1,2 @@
 <b><%= character.name %></b> <%= I18n.t('services.bot_context.representers.check.fate.skill', attr: Fate::Character.skills.dig(target, 'name', I18n.locale.to_s)) %>
-<%= I18n.t('services.bot_context.representers.check.result', result: dnd_roll_result_message(result)) %>
+<%= I18n.t('services.bot_context.representers.check.result') %><%= dnd_roll_result_message(result) %>
