+++ app/views/bots_v2/character_bot/check/nimble/initiative.text.erb
