+++ app/views/bots_v2/character_bot/check/nimble/skill.text.erb
@@ -1,0 +1,2 @@
+<b><%= character.name %></b> <%= I18n.t('services.bot_context.representers.check.dnd.skill', attr: Config.data('nimble', 'skills').dig(target, 'name', I18n.locale.to_s)) %>
+<%= I18n.t('services.bot_context.representers.check.result') %><%= dnd_roll_result_message(result) %>
